@@ -15,10 +15,14 @@ plugins {
 repositories {
     mavenCentral()
 }
+compileJava
+compileJava {
+    options.compilerArgs << '-parameters'
+}
 
 java {
     toolchain {
-        languageVersion = JavaLanguageVersion.of(11)
+        languageVersion = JavaLanguageVersion.of(17)
     }
 
     withJavadocJar()
